honban2の最新版

+++ etrobo_honban2/CompetitionRunning/RunningCalculation.h
@@ -13,7 +13,7 @@ public:
 
 	RunningCalculation(SectionRunningDataGet* SectionRunningDataGet,
 			PIDCalculation* PIDCalculation);
-	int8_t RunningCalculate(DeviceValue dv,int now_section);
+	double RunningCalculate(DeviceValue dv,int now_section);
 
 
 private:
