+++ etrobo_honban2/app.cpp
@@ -190,7 +190,7 @@ void main_task(intptr_t unused) {
 		if (ev3_button_is_pressed(BACK_BUTTON))
 			break;
 
-		tail_control(TAIL_ANGLE_DRIVE); /* バランス走行用角度に制御 */
+		tail_control(TAIL_ANGLE_DRIVE); /*a バランス走行用角度に制御 */
 
 
 		g_unBrightness  = colorSensor->getBrightness();
@@ -203,7 +203,7 @@ void main_task(intptr_t unused) {
 //		}
 
 
-		int8_t x = runningcontroller->RunningExecute(now_section);
+		double x = runningcontroller->RunningExecute(now_section);
 
 		// ログ
 //		char cBuff[1024];
@@ -211,7 +211,7 @@ void main_task(intptr_t unused) {
 //		fputs(cBuff, bt); // エコーバック
 
 	char cBuff[1024];
-	sprintf(cBuff, "Main+++,%d\n",  x);
+	sprintf(cBuff, "Main+++,%f\n",  x);
 	fputs(cBuff, bt); // エコーバック
